Refactored cloud steps into separate files
