+++ www/src/components/shell/cloud/provider.js
@@ -1,20 +1,20 @@
-import { createElement, useCallback, useContext, useRef, useState } from 'react'
-import { Down } from 'grommet-icons'
-import { Box, Drop } from 'grommet'
-import { Div, Flex, H2, Img, P, RadioGroup, Text } from 'honorable'
+import { useCallback, useContext, useState } from 'react'
+import { Box } from 'grommet'
+import { Div, H2, P, RadioGroup, Text } from 'honorable'
 
-import { Button, CloudIcon, Radio } from 'pluralsh-design-system'
+import { Radio } from 'pluralsh-design-system'
 
 import { Provider } from '../../repos/misc'
 
-import { CLOUDS, SECTION_INSTALL_CLI } from '../constants'
-import { CardButton, CreateShellContext, DemoCard, Header, NavSection } from '../CloudShell'
+import { CardButton, CreateShellContext } from '../CloudShell'
 
 import { AWS_VALIDATIONS, AwsForm, awsSynopsis } from './aws'
 import { GCP_VALIDATIONS, GcpForm, gcpSynopsis } from './gcp'
 import { DemoProject } from './demo'
+import { CloudDecision } from './CloudDecision'
+import { CloudCredentials } from './CloudCredentials'
 
-function CloudItem({ provider, setProvider }) {
+export function CloudItem({ provider, setProvider }) {
   return (
     <Box
       direction="row"
@@ -37,7 +37,7 @@ function CloudItem({ provider, setProvider }) {
   )
 }
 
-const ProviderForms = {
+export const ProviderForms = {
   AWS: AwsForm,
   GCP: GcpForm,
 }
@@ -56,7 +56,7 @@ export const synopsis = ({ provider, ...rest }) => {
   }
 }
 
-function CloudOption({ providerLogo, header, description, selected, ...props }) {  
+export function CloudOption({ providerLogo, header, description, selected, ...props }) {  
   return (
     <CardButton
       position="relative"
@@ -88,7 +88,7 @@ function CloudOption({ providerLogo, header, description, selected, ...props }) 
   )
 }
 
-function ChooseAShell({ options, selected, setSelected }) {
+export function ChooseAShell({ options, selected, setSelected }) {
 
   return (
     <Div
@@ -127,103 +127,14 @@ function ChooseAShell({ options, selected, setSelected }) {
   )
 }
 
-function CloudDecision({ doSetPath }) {
-  const { previous, setSection } = useContext(CreateShellContext)
-
-  const [nextPath, setNextPath] = useState()
-  const [byocShell, setByocShell] = useState('cloud')
-
-  return (
-    <>
-      <DemoCard title="Choose a cloud">
-        <P
-          body1
-          color="text-light"
-          marginBottom="medium"
-        >
-          Plural makes it easy to plug into your own cloud, but we also provide a free demo cloud to help you get started.
-        </P> 
-        <Flex mx={-1}>
-          <CloudOption
-            selected={nextPath === 'demo'}
-            providerLogo={(
-              <Img
-                src="/gcp.png"
-                alt="Google Cloud logo"
-                width="100%"
-              />
-            )}
-            header="GCP Cloud Demo"
-            description="A six-hour instance of a GCP cloud to help get you started."
-            onClick={() => setNextPath('demo')}
-          />
-          <CloudOption
-            selected={nextPath === 'byoc'}
-            providerLogo={(
-              <CloudIcon
-                size={40}
-                color="text-light"
-              />
-            )}
-            header="Use Your Own Cloud"
-            description="Plug in your cloud credentials and start building."
-            onClick={() => setNextPath('byoc')}
-          />
-        </Flex>
-        {nextPath === 'byoc' && (
-          <ChooseAShell
-            selected={byocShell}
-            setSelected={setByocShell}
-            options={[{
-              value: 'cloud',
-              label: 'Our cloud shell (quickest)',
-            },
-            {
-              value: 'cli', 
-              label: 'Install the Plural CLI on your local machine',
-            }]}
-          />
-        )}
-      </DemoCard>
-      {/* Navigation */}
-      <NavSection>
-        <Button
-          secondary
-          onClick={() => {
-            previous()
-          }}
-        >
-          Back
-        </Button>
-        <Button
-          disabled={!nextPath}
-          onClick={() => {
-            if (nextPath === 'byoc' && byocShell === 'cli') {
-              setSection(SECTION_INSTALL_CLI)
-            }
-            else {
-              doSetPath(nextPath)
-            } 
-          }}
-        >Continue
-        </Button>
-      </NavSection>
-    </>
-  )
-}
-
 export function ProviderForm() {
-  const { provider, setProvider, workspace, setWorkspace, credentials, setCredentials, demo, setDemo, next } = useContext(CreateShellContext)
-  const ref = useRef()
+  const { setProvider, workspace, setWorkspace, credentials, setCredentials, demo, setDemo, next } = useContext(CreateShellContext)
   const [path, setPath] = useState(null)
-  const [open, setOpen] = useState(false)
-  const close = useCallback(() => setOpen(false), [setOpen])
+
   const doSetPath = useCallback(path => {
     if (path === 'demo') setDemo(true)
     setPath(path)
   }, [setPath, setDemo])
-
-  const form = ProviderForms[provider]
 
   if (!path) {
     return (
@@ -247,55 +158,7 @@ export function ProviderForm() {
   }
 
   return (
-    <>
-      <Box
-        fill
-        align="center"
-        justify="center"
-        gap="small"
-      >
-        <Header text="Cloud Credentials" />
-        <Box
-          ref={ref}
-          direction="row"
-          align="center"
-          gap="small"
-          onClick={() => setOpen(true)}
-          hoverIndicator="card"
-          round="xsmall"
-          pad="small"
-        >
-          <Text
-            size="small"
-            weight={500}
-          >
-            Provider
-          </Text>
-          <Provider
-            provider={provider}
-            width={40}
-          />
-          <Down size="small" />
-        </Box>
-        {createElement(form, { workspace, setWorkspace, credentials, setCredentials })}
-      </Box>
-      {open && (
-        <Drop
-          target={ref.current}
-          onClickOutside={close}
-          onEsc={close}
-        >
-          <Box width="250px">
-            {CLOUDS.map(cloud => (
-              <CloudItem
-                provider={cloud}
-                setProvider={setProvider}
-              />
-            ))}
-          </Box>
-        </Drop>
-      )}
-    </>
+    <CloudCredentials />
   )
-
 }
+
